backend/requirements.txt
autopep8==2.0.0
cachetools==5.2.0
certifi==2022.12.7
charset-normalizer==2.1.1
click==8.1.3
colorama==0.4.6
gensim==4.3.2
google-api-core==2.11.0
google-auth==2.15.0
google-cloud-core==2.3.2
google-cloud-storage==2.7.0
google-crc32c==1.5.0
google-resumable-media==2.4.0
googleapis-common-protos==1.57.0
idna==3.4
joblib==1.2.0
Levenshtein==0.21.1
lightgbm==4.1.0
mpstemmer @ git+https://github.com/ariaghora/mpstemmer.git@25a5fd923af163a7eac3a5ec976984156ca8fa8b
nltk==3.7
protobuf==4.21.10
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.10.0
regex==2022.10.31
requests==2.28.1
rsa==4.9
Sastrawi==1.0.1
scikit-learn==1.3.2
scipy==1.11.2
six==1.16.0
tomli==2.0.1
urllib3==1.26.13